Vulkan: support indices in client memory.
glDrawElements will now work with indices and/or vertex data in client memory,
as well as in a buffer object.
Enable corresponding tests.

BUG=angleproject:1683

+++ src/libANGLE/renderer/vulkan/VertexArrayVk.h
@@ -28,8 +28,8 @@ class VertexArrayVk : public VertexArrayImpl
 
     gl::Error streamVertexData(ContextVk *context,
                                StreamingBuffer *stream,
-                               int firstVertex,
-                               int lastVertex);
+                               size_t firstVertex,
+                               size_t lastVertex);
     void syncState(const gl::Context *context,
                    const gl::VertexArray::DirtyBits &dirtyBits) override;
